Add block IoU as matching confidence proxy
Add matching confidence column when merging attributes to indicate the
reliability of the match that is used to merge the attributes.

The previously existing <attribute>_confidence_iou columns differ from
the matching confidence in that they only consider matched buildings
that have the attribute, but not the quality of the underlying match
itself. To some degree, the matching confidence can be understood as an
upper bound for the attribute confidence.