Add confidence measure for merging attributes between m:n matches
For matches that cannot be adaquately represented as 1:n matches, the
intersection-over-union of each building unit with all its matches is
not well suited to describe the overall similarity of the building
representations betwenn which attributes are merged. A more robust
measure to is the intersection-over-union between the aggregated
matching m:n geometries. The matching pairs are aggregated to m:n
matches by identifying connected components in a bipartite graph.

. . . . . . . . . . . . . . . . . . . . . . . . . . . . .
.               .                       .               .
.   +---------------------------------------------+     .
.   |           .          |            .         |     .
.   |           .          |            .         |     .
.   |           .          |            .         |     .
.   |           .          |            .         |     .
.   +----------------------+----------------------+     .
.               .                       .               .
. . . . . . . . . . . . . . . . . . . . . . . . . . . . .